Re-organize the backend structure
