+++ src/main/java/ch/heigvd/pro/b04/answers/AnswerController.java
@@ -1,8 +1,5 @@
-package ch.heigvd.pro.b04.endpoints;
+package ch.heigvd.pro.b04.answers;
 
-import ch.heigvd.pro.b04.answers.Answer;
-import ch.heigvd.pro.b04.answers.AnswerIdentifier;
-import ch.heigvd.pro.b04.answers.AnswerRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
