+++ src/main/java/ch/heigvd/pro/b04/auth/LoginController.java
@@ -1,9 +1,6 @@
-package ch.heigvd.pro.b04.endpoints;
+package ch.heigvd.pro.b04.auth;
 
-import ch.heigvd.pro.b04.login.TokenCredentials;
-import ch.heigvd.pro.b04.login.UserCredentials;
-import ch.heigvd.pro.b04.login.exceptions.DuplicateUsernameException;
-import ch.heigvd.pro.b04.login.exceptions.UnknownUserCredentialsException;
+import ch.heigvd.pro.b04.auth.exceptions.UnknownUserCredentialsException;
 import ch.heigvd.pro.b04.moderators.Moderator;
 import ch.heigvd.pro.b04.moderators.ModeratorRepository;
 import java.util.Optional;
@@ -44,26 +41,5 @@ public class LoginController {
       }
     });
     return response.orElseThrow(UnknownUserCredentialsException::new);
-  }
-
-  /**
-   * Registers a new user, assuming we got some user credentials.
-   *
-   * @param credentials The credentials to use for registration.
-   * @return An authentication token for the provided account.
-   */
-  @RequestMapping(value = "register", method = RequestMethod.POST)
-  public TokenCredentials register(@RequestBody UserCredentials credentials)
-      throws DuplicateUsernameException {
-    Moderator moderator = Moderator.builder()
-        .username(credentials.getUsername())
-        .secret(credentials.getPassword())
-        .build();
-    try {
-      moderators.saveAndFlush(moderator);
-    } catch (Throwable t) {
-      throw new DuplicateUsernameException();
-    }
-    return TokenCredentials.builder().token(credentials.getPassword()).build();
   }
 }
