+++ src/main/java/ch/heigvd/pro/b04/auth/UserCredentialsSerializer.java
@@ -1,4 +1,4 @@
-package ch.heigvd.pro.b04.login;
+package ch.heigvd.pro.b04.auth;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
