+++ src/main/java/ch/heigvd/pro/b04/auth/exceptions/DuplicateUsernameException.java
@@ -1,4 +1,4 @@
-package ch.heigvd.pro.b04.login.exceptions;
+package ch.heigvd.pro.b04.auth.exceptions;
 
 public class DuplicateUsernameException extends Exception {
