+++ src/main/java/ch/heigvd/pro/b04/auth/exceptions/UnknownUserCredentialsException.java
@@ -1,4 +1,4 @@
-package ch.heigvd.pro.b04.login.exceptions;
+package ch.heigvd.pro.b04.auth.exceptions;
 
 public class UnknownUserCredentialsException extends RuntimeException {
