+++ src/main/java/ch/heigvd/pro/b04/moderators/ModeratorController.java
@@ -1,7 +1,5 @@
-package ch.heigvd.pro.b04.endpoints;
+package ch.heigvd.pro.b04.moderators;
 
-import ch.heigvd.pro.b04.moderators.Moderator;
-import ch.heigvd.pro.b04.moderators.ModeratorRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
